Capston design review:
The design was updated to meet the requirement requested by the
reviewer.

App uses a Loader to move its data to its views.
- It was detailed that the system will use cursor loader to show the
list of golf fields and scorecards.

Must implement at least one of the three
If it regularly pulls or sends data to/from a web service or API, app
updates data in its cache at regular intervals using a SyncAdapter or
Firebase JobDispatcher.
OR
If it needs to pull or send data to/from a web service or API only once,
or on a per request basis (such as a search application), app uses an
IntentService to do so.
OR
It it performs short duration, on-demand requests(such as search), app
uses an AsyncTask.

- it was added a new functionality which notifies the user when he/she
got a best score (Notification will be trgger by intentServices